+++ packages/api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @validation-os/api
 
+## 0.16.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @validation-os/core@0.16.2
+
 ## 0.16.1
 
 ### Patch Changes